Refactor subscription dispatch
The dispatch towards subscriptions needs to be split out into the usual
two steps, prepare and execute.

This patch makes some preparatory steps:
- prepare() can result in a PreparedRequest
- even HEAD requests need to parse the proposed subscription ID
- bad subscription IDs should not report Bad Request, as the request can
  very much be valid, it's just that if the string does not parse to
  Uint32, we know there is no such subscription
- ReceiverHolder's subscriptionId() is now of proper type

JIRA: NETCONF-1449

+++ plugins/restconf-server-mdsal/src/main/java/org/opendaylight/restconf/server/mdsal/MdsalRestconfStreamRegistry.java
@@ -227,7 +227,7 @@ public final class MdsalRestconfStreamRegistry extends AbstractRestconfStreamReg
         tx.merge(LogicalDatastoreType.OPERATIONAL, YangInstanceIdentifier.of(
             SUBSCRIPTIONS_NODEID,
             SUBSCRIPTION_NODEID,
-            subscriptionArg(Uint32.valueOf(receiver.subscriptionId())),
+            subscriptionArg(receiver.subscriptionId()),
             RECEIVERS_NODEID,
             RECEIVER_NODEID,
             receiverArg(receiver.receiverName()),
